requirements.txt
Django>=4.2
ultralytics>=8.0.0
pillow
python-dotenv
pyserial
opencv-python
djangorestframework
